+++ addons/binding/org.openhab.binding.loxone/src/main/java/org/openhab/binding/loxone/internal/core/LxServer.java
@@ -685,7 +685,7 @@ public class LxServer {
         if (control != null) {
             control.update(json, room, category);
         } else {
-            control = LxControl.createControl(socketClient, id, json, room, category);
+            control = LxControlFactory.createControl(socketClient, id, json, room, category);
         }
         if (control != null) {
             updateControls(control);
